fix(groups): implement aggregation table.
affects: webiny-api-cms, webiny-api

Previously, IDs of groups would be saved inline in table, in column 'groups'. This is not convenient
when there is a need if a group is being used, eg. before making a delete. That's why, a separate
aggregation table is made, which holds links to entities and groups.