fix: move latestVersion handling into beforeCreate hook
The problem here is that hooks are executed in a very specific order: beforeSave, beforeCreate, afterSave, afterCreate. This particular order caused index generation to be executed before the "latestVersion" flag was set to "false" for the previous revision and so the Search catalog contained the wrong data. By unsetting the "latestVersion" flag in the "beforeCreate", indexes are now generated based on the correct state of entries.