fix(app-plugin-security-cognito-theme): remove built-in password validator
This will not validate password pattern on the client side. Instead, it will let Authentication Provider plugin on the API handle it and throw error if password is invalid. We can somehow fetch and generate password validator in the future, but that would require fetching a password policy from the Authentication Provider itself, and generating a validator based on those rules.