fix: add editor.js bundle to RTE component
This is necessary to apply a fix for occasional block selection being -1, when multiple editors are mounted at the same time. This causes everything to crash. Will open an issue with a repro case for codex team to review, but in the meantime we have to include the patched bundle in our code.