test: update test (create test handler with `console.log` call)
Instead of adding the `console.log` call in the actual application code, with this test, we're doing it just locally. We're creating a new test handler, and in it, a new route handler that just does a simple `console.log(forwardUrl)` call, which we can then inspect in the test.